+++ app.js
@@ -10,12 +10,12 @@ const passport = require("passport");
 const GoogleStrategy = require("passport-google-oauth20").Strategy;
 // const ensureLoggedIn = require("connect-ensure-login").ensureLoggedIn;
 
-const getDataUrl = require("./util/qrcodeUtil");
-const { decrypt } = require("./util/cryptoUtil");
+// const getDataUrl = require("./util/qrcodeUtil");
+// const { decrypt } = require("./util/cryptoUtil");
 const Attend = require("./store/Attend.js");
 
-const rootRouter = require("./route/root");
-// const attendRouter = require("./route/attend");
+const rootRouter = require("./route/rootRouter");
+const attendRouter = require("./route/attendRouter");
 // const eventRouter = require("./route/event");
 
 passport.use(
@@ -64,6 +64,8 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 app.use("/", rootRouter);
+//TODO: /attends로 변경해야함.
+app.use("/register", attendRouter);
 // app.use("/events", eventsRouter);
 
 /*
@@ -105,10 +107,10 @@ app.get("/reader", function (req, res) {
   res.render("reader");
 });
 
-app.get("/Ilist", async function (req, res) {
-  const list = await Attend.findAll();
-  res.render("list", { list });
-});
+// app.get("/Ilist", async function (req, res) {
+//   const list = await Attend.findAll();
+//   res.render("list", { list });
+// });
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
